celery/django/requirements.txt
django
django-celery[redis]
redis
geoip2
mysqlclient
Pillow
pysnmp
pygeoip
sorl-thumbnail
rrdtool
easysnmp
djangorestframework
